requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
fpdf
langchain_google_genai
faiss-cpu
langchain-community
fitz
pdf2image
pymupdf
langchain-Huggingface
matplotlib
seaborn